refactor(): package를 feature를 기준으로 정리
