+++ src/main/java/young/board/comment/CommentController.java
@@ -1,10 +1,9 @@
-package young.board.controller;
+package young.board.comment;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import young.board.service.CommentService;
 
 @RestController
 @RequiredArgsConstructor
