+++ src/main/java/young/board/post/PostService.java
@@ -1,11 +1,11 @@
-package young.board.service;
+package young.board.post;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import young.board.domain.Category;
 import young.board.domain.Post;
-import young.board.repository.post.PostRepository;
+import young.board.post.repository.PostRepository;
 
 import java.util.List;
 import java.util.Optional;
