+++ src/main/java/young/board/recommendation/RecommendationController.java
@@ -1,4 +1,4 @@
-package young.board.controller;
+package young.board.recommendation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -8,7 +8,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import young.board.service.RecommendationService;
 
 @Controller
 @RequiredArgsConstructor
