+++ src/main/java/young/board/recommendation/repository/H2RecommendationRepository.java
@@ -1,4 +1,4 @@
-package young.board.repository.recommendation;
+package young.board.recommendation.repository;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
