+++ src/test/java/young/board/repository/comment/H2CommentRepositoryTest.java
@@ -1,27 +1,25 @@
 package young.board.repository.comment;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
+import young.board.comment.repository.CommentRepository;
 import young.board.domain.Category;
 import young.board.domain.Comment;
 import young.board.domain.Post;
-import young.board.repository.post.PostRepository;
-
-import javax.persistence.EntityManager;
+import young.board.post.repository.PostRepository;
 
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
 class H2CommentRepositoryTest {
-    @Autowired CommentRepository commentRepository;
+    @Autowired
+    CommentRepository commentRepository;
     @Autowired PostRepository postRepository;
 
     Post post1;
